requirements.txt
Flask
pymongo
gunicorn
python-dotenv
bcrypt
pandas
numpy
requests
elasticsearch==8.11.0